pyproject.toml
[tool.poetry]
name          = "broken-source"
version       = "2024.3.24"
description   = "Broken Source Software Framework"
license       = "AGPL-3.0-only"
authors       = ["Tremeschin <tremeschin@users.noreply.github.com>"]
readme        = "Readme.md"
homepage      = "https://github.com/BrokenSource/BrokenSource"
repository    = "https://github.com/BrokenSource/BrokenSource"
documentation = "https://github.com/BrokenSource/BrokenSource"

[[tool.poetry.packages]]
include = "Broken"

[tool.poetry.scripts]
broken = "Broken.__main__:main"

[tool.poetry.dependencies]
python         = ">=3.10,<3.13"
rich           = "^13.3.4"
dotmap         = "^1.3.30"
arrow          = "^1.2.3"
datetime       = "^5.1"
distro         = "^1.8.0"
forbiddenfruit = "^0.1.4"
requests       = "^2.28.2"
toml           = "^0.10.2"
appdirs        = "^1.4.4"
poetry         = "^1.5.1"
tqdm           = "^4.65.0"
attrs          = "^23.1.0"
numpy          = "^1.25.2"
typer          = "^0.9.0"
intervaltree   = "^3.1.0"
python-dotenv  = "^1.0.0"
pyfiglet       = "^1.0.2"
pyyaml         = "^6.0.1"
imageio        = "^2.33.1"
validators     = "^0.22.0"
requests-cache = "^1.1.1"
pretty-errors  = "^1.2.25"
cachetools     = "^5.3.2"
pyzmq          = "^25.1.2"
pillow         = "^10.0.1"
yaspin         = "^3.0.1"
pyturbojpeg    = "^1.7.3"
opencv-contrib-python = "^4.9.0.80"

# Todo: Gotta use 'em
watchdog       = "^3.0.0"
schedule       = "^1.2.0"
thefuzz        = "^0.19.0" # Maybe rapidfuzz ?
cattrs         = "^23.2.3"
ujson          = "^5.9.0"

[tool.poetry.group.dev.dependencies]
isort       = "^5.12.0"
snakeviz    = "^2.2.0"
pyshortcuts = "^1.9.0"
userpath    = "^1.9.1"
pytest      = "^7.4.4"
meson       = "^1.3.1"
ninja       = "^1.11.1.1"
wheel       = "^0.42.0"
poethepoet  = "^0.24.4"
tuna        = "^0.5.11"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocstrings = {extras = ["python"], version = "^0.24.1"}
mkdocs-gen-files = "^0.5.0"
mkdocs-include-markdown-plugin = "^6.0.4"
mkdocs-autorefs = "^1.0.1"
mkdocs-literate-nav = "^0.6.1"
mkdocs-section-index = "^0.3.8"
mkdocs-material = "^9.5.13"

[tool.poetry.group.staging.dependencies]
redis   = "^5.0.1"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts      = "-ra --no-header --durations=0 -v -p no:cacheprovider"
python_files = "**/__test__.py"
